+++ util/env_posix.cc
@@ -42,6 +42,7 @@
 #include "util/random.h"
 #include "util/iostats_context_imp.h"
 #include "util/rate_limiter.h"
+#include "util/thread_status_impl.h"
 
 // Get nano time for mach systems
 #ifdef __MACH__
@@ -74,6 +75,10 @@
 int rocksdb_kill_odds = 0;
 
 namespace rocksdb {
+
+#if ROCKSDB_USING_THREAD_STATUS
+extern ThreadStatusImpl thread_local_status;
+#endif
 
 namespace {
 
@@ -1570,6 +1575,17 @@ class PosixEnv : public Env {
       return static_cast<int>(thread_id) >= total_threads_limit_;
     }
 
+    // Return the thread priority.
+    // This would allow its member-thread to know its priority.
+    Env::Priority GetThreadPriority() {
+      return priority_;
+    }
+
+    // Set the thread priority.
+    void SetThreadPriority(Env::Priority priority) {
+      priority_ = priority;
+    }
+
     void BGThread(size_t thread_id) {
       bool low_io_priority = false;
       while (true) {
@@ -1651,8 +1667,14 @@ class PosixEnv : public Env {
       BGThreadMetadata* meta = reinterpret_cast<BGThreadMetadata*>(arg);
       size_t thread_id = meta->thread_id_;
       ThreadPool* tp = meta->thread_pool_;
+      // for thread-status
+      thread_local_status.SetThreadType(
+          (tp->GetThreadPriority() == Env::Priority::HIGH ?
+              ThreadStatus::ThreadType::ROCKSDB_HIGH_PRIORITY :
+              ThreadStatus::ThreadType::ROCKSDB_LOW_PRIORITY));
       delete meta;
       tp->BGThread(thread_id);
+      thread_local_status.UnregisterThread();
       return nullptr;
     }
 
@@ -1753,6 +1775,7 @@ class PosixEnv : public Env {
     std::atomic_uint queue_len_;  // Queue length. Used for stats reporting
     bool exit_all_threads_;
     bool low_io_priority_;
+    Env::Priority priority_;
   };
 
   std::vector<ThreadPool> thread_pools_;
@@ -1767,6 +1790,10 @@ PosixEnv::PosixEnv() : checkedDiskForMmap_(false),
                        page_size_(getpagesize()),
                        thread_pools_(Priority::TOTAL) {
   PthreadCall("mutex_init", pthread_mutex_init(&mu_, nullptr));
+  for (int pool_id = 0; pool_id < Env::Priority::TOTAL; ++pool_id) {
+    thread_pools_[pool_id].SetThreadPriority(
+        static_cast<Env::Priority>(pool_id));
+  }
 }
 
 void PosixEnv::Schedule(void (*function)(void*), void* arg, Priority pri) {